backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt
ffmpeg-python
sendgrid
email-validator
python-dotenv
Pillow